digital-twin-mvp/app/requirements.txt.txt
paho-mqtt
pandas
python-dotenv
streamlit
matplotlib
numpy
pytest
